// src/types/CopyTradeMasterOrderHistory.ts
// Generated by https://quicktype.io

export default interface CopyTradeMasterOrderHistory {

    // FIELDS FROM API
    count: number | null;
    next: number | null;
    previous: number | null;
    results: CopyTradeMasterOrderHistoryResult[];
}

export interface CopyTradeMasterOrderHistoryResult {
    // CUSTOM FIELDS
    checkbox?: string;
    copy_trade_follower_amount?: string;

    // FIELDS FROM API
    id: number;
    ss_id: number;
    o_fold: number;
    o_type: string;
    o_time: number;
    total_followtrade: number;
    total_follow: number;
    o_status: string;
    o_result: number;
    o_owner: number;
    owner_name: string;
    created_at: string;
    updated_at: string;
}